requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-genai==0.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
firebase-admin==6.4.0
